🔨 [TEA-62]: Refactor api endpoints into struct methods
All endpoints are now part of api.API, a struct with a variable for the dbPool, that can be easily instantiated from main.go and utilized in router.go without using lots of redundant code and variable passing.
The db connection established by individual endpoints now has its own method, API.dbConn(ctx), which handles the aquisition of the connection and the error handling (complete with http response status code 500) should that go awry.
This allows endpoints' code to be cleaner, more legible, and DRY. More maintainable in the long run!

The code in router.go also got split into an additional function that handles the mounting of individual api endpoints so that function line limits won't continue to be a concern as we expand.

Also includes some cleanup, mainly removal of example endpoint greeter.go and old comments